requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==2.2.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
distro==1.9.0
executing==1.2.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.5.1
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.2
idna==3.10
ipykernel==6.29.5
ipython==8.15.0
jedi==0.18.2
Jinja2==3.1.6
jiter==0.10.0
jupyter_client==8.3.0
jupyter_core==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.4.2
numpy==2.2.6
openai==1.82.0
packaging==25.0
parso==0.8.3
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.15.1
python-dotenv==1.1.0
python-multipart==0.0.20
pywin32==305.1
PyYAML==6.0.2
pyzmq==25.1.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==80.9.0
sniffio==1.3.1
stack-data==0.6.2
starlette==0.46.2
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
tornado==6.3.1
tqdm==4.67.1
traitlets==5.9.0
transformers==4.52.3
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.6
